# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AeroNode API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces the current user belongs to
      responses:
        "200":
          description: Workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceSummary"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceSummary"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace detail including members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetail"
        "403":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update workspace name/description (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetail"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a workspace (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /workspaces/{id}/members:
    post:
      operationId: addWorkspaceMember
      tags: [workspaces]
      summary: Add a registered user to the workspace by email (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMember"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "409":
          description: Already a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /workspaces/{id}/members/{userId}:
    patch:
      operationId: updateMemberRole
      tags: [workspaces]
      summary: Change a member's role (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMember"
    delete:
      operationId: removeMember
      tags: [workspaces]
      summary: Remove a member from the workspace (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /workspaces/{id}/activity:
    get:
      operationId: listActivity
      tags: [workspaces]
      summary: List the 100 most recent activity log entries for a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity entries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEntry"
  /workspaces/{id}/export:
    get:
      operationId: exportWorkspace
      tags: [workspaces]
      summary: Export a workspace's full data as JSON
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace export bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceExport"
...
components:
  schemas:
...
    ErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Role:
      type: string
      enum: [owner, editor, viewer]
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
...
    WorkspaceInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
      required: [name]
    WorkspaceUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
    WorkspaceSummary:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        role:
          $ref: "#/components/schemas/Role"
        memberCount: { type: integer }
        taskCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        [id, name, description, role, memberCount, taskCount, createdAt, updatedAt]
    WorkspaceMember:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        name: { type: string }
        email: { type: string }
        avatarColor: { type: string }
        role:
          $ref: "#/components/schemas/Role"
        joinedAt: { type: string, format: date-time }
      required: [id, userId, name, email, avatarColor, role, joinedAt]
    WorkspaceDetail:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        myRole:
          $ref: "#/components/schemas/Role"
        members:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceMember"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, description, myRole, members, createdAt, updatedAt]
    MemberInput:
      type: object
      properties:
        email: { type: string }
        role:
          $ref: "#/components/schemas/Role"
      required: [email, role]
    MemberRoleUpdate:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/Role"
      required: [role]
    ActivityEntry:
      type: object
      properties:
        id: { type: integer }
        workspaceId: { type: integer }
        userId: { type: integer }
        userName: { type: string }
        message: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, workspaceId, userId, userName, message, createdAt]
    Task:
      type: object
      properties:
        id: { type: integer }
        workspaceId: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        dueDate: { type: ["string", "null"], format: date-time }
        order: { type: integer }
        assigneeId: { type: ["integer", "null"] }
        assigneeName: { type: ["string", "null"] }
        createdBy: { type: integer }
        lockedBy: { type: ["integer", "null"] }
        lockedByName: { type: ["string", "null"] }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        [id, workspaceId, title, description, status, priority, dueDate, order, assigneeId, assigneeName,
         createdBy, lockedBy, lockedByName, version, createdAt, updatedAt]
...
    WorkspaceExport:
      type: object
      properties:
        workspace:
          $ref: "#/components/schemas/WorkspaceDetail"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        activity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEntry"
        exportedAt: { type: string, format: date-time }
      required: [workspace, tasks, activity, exportedAt]
...