# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AeroNode API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEntry {
  id: number;
  workspaceId: number;
  userId: number;
  userName: string;
  message: string;
  createdAt: Date;
}