# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AeroNode API specification
 * OpenAPI spec version: 0.1.0
 */

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  owner: 'owner',
  editor: 'editor',
  viewer: 'viewer',
} as const;