# lib/api-zod/src/generated/types/workspaceSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AeroNode API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from './role';

export interface WorkspaceSummary {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  role: Role;
  memberCount: number;
  taskCount: number;
  createdAt: Date;
  updatedAt: Date;
}